Getting results from the detections provided by the MOT Challenge. Also, created a script to generate the video with the results of the MPNTrack.

+++ yolov8_bytetrack_motchallenge/run/infer.py
@@ -97,7 +97,7 @@ with VideoSink(
                     z,
                 ) in zip(
                     [frame_idx] * len(detections),
-                    results.boxes.id[mask_classes_interest].cpu().numpy().astype(int),
+                    results.boxes.id[mask_classes_interest].cpu().numpy(),
                     results.boxes.xywh[mask_classes_interest].cpu().numpy(),
                     results.boxes.conf[mask_classes_interest].cpu().numpy(),
                     [-1] * len(detections),
